# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EV Community Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerOwner
      tags: [auth]
      summary: Submit EV owner registration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterOwnerBody"
      responses:
        "201":
          description: Registration submitted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistrationResult"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: loginOwner
      tags: [auth]
      summary: Login with credentials
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResult"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Owner"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── OWNERS ─────────────────────────────────────────────────────────────────
...
  /admin/registrations:
    get:
      operationId: listRegistrations
      tags: [admin]
      summary: List all registrations (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected]
      responses:
        "200":
          description: List of registrations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Registration"
  /admin/registrations/{id}/approve:
    post:
      operationId: approveRegistration
      tags: [admin]
      summary: Approve a registration and set login credentials
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveRegistrationBody"
      responses:
        "200":
          description: Registration approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/registrations/{id}/reject:
    post:
      operationId: rejectRegistration
      tags: [admin]
      summary: Reject a registration
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectRegistrationBody"
      responses:
        "200":
          description: Registration rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ─── DASHBOARD ───────────────────────────────────────────────────────────────
...
  /blog/posts:
    get:
      operationId: listBlogPosts
      tags: [blog]
      summary: List blog posts
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: authorId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of blog posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPostList"

    post:
      operationId: createBlogPost
      tags: [blog]
      summary: Create a new blog post (approved owners only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBlogPostBody"
      responses:
        "201":
          description: Blog post created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
  /blog/posts/{id}:
    get:
      operationId: getBlogPost
      tags: [blog]
      summary: Get a specific blog post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
        "404":
          description: Post not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      operationId: updateBlogPost
      tags: [blog]
      summary: Update a blog post (author only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBlogPostBody"
      responses:
        "200":
          description: Updated blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"

    delete:
      operationId: deleteBlogPost
      tags: [blog]
      summary: Delete a blog post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Post deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ─── RANGE CALCULATOR ────────────────────────────────────────────────────────
...
  /accessories:
    get:
      operationId: listAccessories
      tags: [accessories]
      summary: List EV accessories
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: carModel
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of accessories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Accessory"

    post:
      operationId: createAccessory
      tags: [accessories]
      summary: Add an accessory (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccessoryBody"
      responses:
        "201":
          description: Accessory created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Accessory"
  /accessories/{id}:
    patch:
      operationId: updateAccessory
      tags: [accessories]
      summary: Update an accessory (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAccessoryBody"
      responses:
        "200":
          description: Accessory updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Accessory"

    delete:
      operationId: deleteAccessory
      tags: [accessories]
      summary: Delete an accessory (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Accessory deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ─── GALLERY ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # ─── AUTH ─────────────────────────────────────────────────────────────────
    RegisterOwnerBody:
      type: object
      properties:
        email:
          type: string
          format: email
        name:
          type: string
        age:
          type: integer
          minimum: 16
          maximum: 100
        phoneNumber:
          type: string
        whatsappNumber:
          type: string
        telegramId:
          type: string
        areaOfStay:
          type: string
        occupation:
          type: string
        vehicleNumber:
          type: string
        evCarsOwned:
          type: array
          items:
            type: string
        variantColor:
          type: string
        otherCarModel:
          type: string
        purchaseMonthYear:
          type: string
        proofOfOwnershipUrl:
          type: string
      required: [email, name, age, phoneNumber, whatsappNumber, telegramId, areaOfStay, occupation, vehicleNumber, evCarsOwned, variantColor, purchaseMonthYear]
    RegistrationResult:
      type: object
      properties:
        id:
          type: integer
        message:
          type: string
        status:
          type: string
      required: [id, message, status]
    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
    LoginResult:
      type: object
      properties:
        owner:
          $ref: "#/components/schemas/Owner"
        token:
          type: string
      required: [owner, token]

    # ─── OWNER ───────────────────────────────────────────────────────────────
    Owner:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        age:
          type: integer
        phoneNumber:
          type: string
        whatsappNumber:
          type: string
        telegramId:
          type: string
        areaOfStay:
          type: string
        city:
          type: string
        occupation:
          type: string
        vehicleNumber:
          type: string
        evCarsOwned:
          type: array
          items:
            type: string
        variantColor:
          type: string
        otherCarModel:
          type: string
        purchaseMonthYear:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        role:
          type: string
          enum: [user, admin]
        blogEnabled:
          type: boolean
        username:
          type: string
        avatarUrl:
          type: string
        bio:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, email, name, status, role]
...
    Registration:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        age:
          type: integer
        phoneNumber:
          type: string
        whatsappNumber:
          type: string
        telegramId:
          type: string
        areaOfStay:
          type: string
        occupation:
          type: string
        vehicleNumber:
          type: string
        evCarsOwned:
          type: array
          items:
            type: string
        variantColor:
          type: string
        otherCarModel:
          type: string
        purchaseMonthYear:
          type: string
        proofOfOwnershipUrl:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        rejectionReason:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, email, name, status]
    ApproveRegistrationBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        blogEnabled:
          type: boolean
      required: [username, password]
    RejectRegistrationBody:
      type: object
      properties:
        reason:
          type: string
      required: [reason]

    # ─── DASHBOARD ────────────────────────────────────────────────────────────
...
    BlogPost:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        excerpt:
          type: string
        coverImageUrl:
          type: string
        tags:
          type: array
          items:
            type: string
        authorId:
          type: integer
        authorName:
          type: string
        authorAvatarUrl:
          type: string
        published:
          type: boolean
        viewCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, title, content, authorId, authorName, published, createdAt]
    BlogPostList:
      type: object
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/BlogPost"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [posts, total, page, limit]
    CreateBlogPostBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        excerpt:
          type: string
        coverImageUrl:
          type: string
        tags:
          type: array
          items:
            type: string
        published:
          type: boolean
      required: [title, content]
    UpdateBlogPostBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        excerpt:
          type: string
        coverImageUrl:
          type: string
        tags:
          type: array
          items:
            type: string
        published:
          type: boolean

    # ─── RANGE ────────────────────────────────────────────────────────────────
...
    Accessory:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        description:
          type: string
        compatibleCars:
          type: array
          items:
            type: string
        sellerName:
          type: string
        sellerContact:
          type: string
        sellerUrl:
          type: string
        location:
          type: string
        priceRange:
          type: string
        imageUrl:
          type: string
        rating:
          type: number
        createdAt:
          type: string
          format: date-time
      required: [id, name, category, sellerName]
    CreateAccessoryBody:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        description:
          type: string
        compatibleCars:
          type: array
          items:
            type: string
        sellerName:
          type: string
        sellerContact:
          type: string
        sellerUrl:
          type: string
        location:
          type: string
        priceRange:
          type: string
        imageUrl:
          type: string
        rating:
          type: number
      required: [name, category, sellerName]
    UpdateAccessoryBody:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        description:
          type: string
        compatibleCars:
          type: array
          items:
            type: string
        sellerName:
          type: string
        sellerContact:
          type: string
        sellerUrl:
          type: string
        location:
          type: string
        priceRange:
          type: string
        imageUrl:
          type: string
        rating:
          type: number

    # ─── GALLERY ──────────────────────────────────────────────────────────────
...